[ReactDnd] Reset to asana's version of react-dnd.d.ts
